% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front_diversity.R
\name{area_btw_front}
\alias{area_btw_front}
\title{area_btw_front}
\usage{
area_btw_front(front_sup, front_inf, sens, n = 10^4)
}
\arguments{
\item{front_sup}{set of n point having p coordinates store in a matrix (n x p).
front_sup is better than front_inf as define by sens}

\item{front_inf}{set of n point having p coordinates store in a matrix (n x p).
front_inf is worst than front_sup as define by sens}

\item{sens}{vector of size p with value in c("max", "min") giving the objectif
seeking for each component.}

\item{n}{number of point sample to proceed to the Monte Carlos estimation}
}
\value{
return a numerical value, being a area for p=2, a volume if p=3, and
so on
}
\description{
Estimate the area between two front of dimension p throught a Monte Carlos
algorithme.
}
