% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front_diversity.R
\name{comparaison}
\alias{comparaison}
\title{comparaison}
\usage{
comparaison(X1, X2, sens)
}
\arguments{
\item{X1}{a numeric vector of size p which will be compare to the second one}

\item{X2}{a numeric vector of size p which will be compare to the first one}

\item{sens}{a character vector of size p allowing only the two value in
c("min", "max"). A value "min" of sens at the j position means to test if
X1 < X2.}
}
\value{
return TRUE if X1 is better than X2 on all the component given the
set of instruction sens, FALSE if it is the opposite and NULL if for all the
component none between X1 and X2 is better.
}
\description{
Function to compare to vector on all their component given a set of
priorities
}
