% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front_diversity.R
\name{count_under_front}
\alias{count_under_front}
\title{count_under_front}
\usage{
count_under_front(front, X, sens)
}
\arguments{
\item{front}{a numerical matrix n x p, n being the number of point forming
the front.}

\item{X}{a numerical matrix n2 x p, n2 being the number of position points
to test}

\item{sens}{a character vector of size p allowing only the two value in
c("min", "max"). A "min" value at the j index of sens means that the function
will count the number of point below the front of the j component and
vice versa.}
}
\value{
return the number of point
}
\description{
Function that count the number of point below/above the front given priorities
}
