% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_operators (SFConflict
%   alexandre.conanec@u-bordeaux.fr 2020-01-17-11-05-42).R,
%   R/genetic_operators.R
\name{crossing_over}
\alias{crossing_over}
\title{Crossing over operator}
\usage{
crossing_over(X, parents, crossing_over_size)

crossing_over(X, parents, crossing_over_size)
}
\arguments{
\item{couple}{a matrix/data.frame of two individus}

\item{crossing_position}{a sorted vector of variable position to switch}

\item{couple}{a matrix/data.frame of two individus}

\item{crossing_position}{a sorted vector of variable position to switch}
}
\value{
a matrix/data.frame of the two childs from the crossing over of the
parents

a matrix/data.frame of the two childs from the crossing over of the
parents
}
\description{
This function switch components for the given variables position between two
individus (parent) which have been selected after the tounament

This function switch components for the given variables position between two
individus (parent) which have been selected after the tounament
}
\examples{
sum(1:10)
sum(1:10)
}
