% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crowding_distance.R
\name{crowding_distance}
\alias{crowding_distance}
\title{Crowding distance}
\usage{
crowding_distance(Y, threshold = 10^(-5) * sapply(Y[, -((NCOL(Y) -
  1):NCOL(Y))], sd))
}
\arguments{
\item{S}{a matrix/data.frame of evaluated points belonging to the same front}
}
\value{
A vector containing the crowding distances
}
\description{
Compute the crowding distance between all the point belonging to a same
front
}
\examples{
sum(1:10)
}
