% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance_ranking.R
\name{dominance_ranking}
\alias{dominance_ranking}
\title{Dominance ranking}
\usage{
dominance_ranking(X, sens)
}
\arguments{
\item{X}{a matrix/data.frame of all the selected individus evaluation}

\item{sens}{vector of size NCOL(X) containing either "min" or "max"
to choose how to optimize each objectif.}
}
\value{
a vecteur of rank associated with the individus of X
}
\description{
Use the pairwisecomparison function to attribute a rank of individu belonging
to the same front which mean that they are not comparable to each other.
The lowest the rank the best is the individu
}
\examples{
sum(1:10)
}
