% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDF.R
\name{F_n}
\alias{F_n}
\title{Cumulative distribution function (cdf)}
\usage{
F_n(X, Y, y, x, h_n = NULL)
}
\arguments{
\item{X}{a numerical vector, matrix, or data.frame of explanatory variables of dimension nxq}

\item{Y}{a numerical vector of response variable of size n}

\item{y}{the value for which we want to evaluate the probability of being
lower or equal to the random variable Y knowing that X=x}

\item{x}{the conditional value of the random variable X}

\item{h_n}{the mobile windows size use to smooth the distribution function.}
}
\value{
P(Y<=y|X=x)
}
\description{
Non parametric estimation of Cumulative distribution function (cdf).
This can be express as P(Y<=y|X=x)
}
\examples{
library(tidyverse)

set.seed(258164)
n <- 300
X <- runif(n = n, min = -2, max = 2)
Y <- X^2 + rnorm(n)

#parametres
h_n <- hopt(X, Y)$h

#estimation
F_n(X = X, Y = Y, x = x, y = y, h_n = h_n)
pnorm(y, mean = x^2, sd = 1) #true probability

}
