% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front_diversity.R
\name{front_diversity}
\alias{front_diversity}
\title{front_diversity}
\usage{
front_diversity(all_front, sens, n = 10^4)
}
\arguments{
\item{all_front}{a numerical matrix n x p, n being the number of point
belonging at all the differents pareto front.}

\item{sens}{vector of size p with value in c("max", "min") giving the objectif
seeking for each component.}

\item{n}{number of point sample to proceed to the Monte Carlos estimation}
}
\value{
return a numerical value, being a area for p=2, a volume if p=3, and
so on, corresponding to the diversity of the fronts
}
\description{
Function to evaluate the diversity of a set of pareto front which aim to
reach the same optimal but which reach different fronts
}
