% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_operators (SFConflict
%   alexandre.conanec@u-bordeaux.fr 2020-01-17-11-05-42).R,
%   R/genetic_operators.R
\name{mutation}
\alias{mutation}
\title{Mutation operator}
\usage{
mutation(Qt, freq, distri_Xi, distribution = "uniform")

mutation(Qt, freq, distri_Xi, distribution = "uniform")
}
\arguments{
\item{Qt}{a matrix/data.frame of all the individus generated after crossing
over}

\item{freq}{a vector of numerics value varying from 0 to 1 which represent
the frequency of mutation of each variable}

\item{distribution}{choice of the distribution where for the mutation value}

\item{Qt}{a matrix/data.frame of all the individus generated after crossing
over}

\item{freq}{a vector of numerics value varying from 0 to 1 which represent
the frequency of mutation of each variable}

\item{distribution}{choice of the distribution where for the mutation value}
}
\value{
a matrix/data.frame of the two childs from the crossing over of the
parents

a matrix/data.frame of the two childs from the crossing over of the
parents
}
\description{
This function assign random value to the individu caracteristique following
a frequency of occurence

This function assign random value to the individu caracteristique following
a frequency of occurence
}
\examples{
sum(1:10)
sum(1:10)
}
