% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_operators.R
\name{mutation_simple}
\alias{mutation_simple}
\title{Mutation operator}
\usage{
mutation_simple(X, freq_m, distri_Xi, type_var)
}
\arguments{
\item{freq_m}{a vector of numerics value varying from 0 to 1 which represent
the frequency of mutation of each variable}

\item{Qt}{a matrix/data.frame of all the individus generated after crossing
over}

\item{distribution}{choice of the distribution where for the mutation value}
}
\value{
a matrix/data.frame of the two childs from the crossing over of the
parents
}
\description{
This function assign random value to the individu caracteristique following
a frequency of occurence
}
\examples{
sum(1:10)
}
