% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsga.R
\name{NSGA}
\alias{NSGA}
\title{NSGA II algorithm}
\usage{
NSGA(X, ff, N, crossing_over_size, freq_mutation, B = 500,
  verbose = TRUE)
}
\arguments{
\item{X}{a matrix/data.frame of an initial population}

\item{ff}{a list of list containing the function or predicting model and
if the objectif function should be maximise "max" or minimize "min"}

\item{N}{the size of the population P before the generation of individus.
By default N=NROW(X)/2}

\item{crossing_over_size}{number of variable to switch when doing the
crossing over}

\item{freq_mutation}{a vector of numerics value varying from 0 to 1 which
represent the frequency of mutation of each variable}
}
\value{
a vecteur of rank associated with the individus of X
}
\description{
The non sorting genetic algorithm is finding the Pareto front
}
\examples{
sum(1:10)
}
