% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optisure.R
\name{optisure}
\alias{optisure}
\title{optisure}
\usage{
optisure(X, Y, sens, alpha = 0.5)
}
\description{
optisure
}
\examples{
# generate data ----
library(tidyverse)
set.seed(1234)
n <- 300

#X quanti
d1 = 3
mini <- rep(x = -2, times = d1)
maxi <- rep(x = 2, times = d1)
X1 <- lapply(seq_len(d1), function(k){
  data.frame(runif(n = n, min = mini[k], max = maxi[k]))
}) \%>\% bind_cols()
colnames(X1) = paste0("X", seq_len(d1))

#X quali
d2 = 2
lev = sample(x = 2:4, d2, replace = TRUE)
X2 <- lapply(seq_len(d2), function(k){
  data.frame(as.factor(sample(x = seq_len(lev[k]), size = n, replace = TRUE)))
}) \%>\% bind_cols()
colnames(X2) = paste0("X", (d1+1):(d1+d2))

X = cbind(X1, X2)
head(X)

#calcul des Y selon une relation lineaire + du bruit
p = 2
ff = lapply(seq_len(p), function(j){
  beta = runif(n = (d1 + sum(lev-1)), min = -2, max = 2)
  function(X) {
    is_num = sapply(X, is.numeric)
    X1 = X[, is_num]
    X2 = X[, !is_num]
    X2 = model.matrix(~., data = X2)[,-1]
    X = cbind(X1, X2)
    as.matrix(X) \%*\% beta
  }
})
names(ff) = paste0("Y", seq_len(p))

Y = lapply(ff, function(f){
  f(X) + rnorm(n)
}) \%>\% as.data.frame(row.names = seq_len(n))

optisure(X, Y, sens = rep("min", p), alpha = 0.5)

plot(res$Y \%>\% arrange(Y1), 'b')

}
