% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparaison.R
\name{pairwise_comparison}
\alias{pairwise_comparison}
\title{Pairwise comparison}
\usage{
pairwise_comparison(X, sens)
}
\arguments{
\item{X}{a matrix/data.frame of all the selected individus evaluation}

\item{sens}{the objectif fonctions goal which can be either "min" or "max"}
}
\value{
a list of two element for each individu. First the dominated_count
which mean how many individu are better. Secondly the dominating_index
which gather the index of the dominted individu.
}
\description{
Compare one by one all the individu evaluations and give them a rank and
count how many individu dominate each individus and the individus that are
dominated by each individus
}
\examples{
sum(1:10)
}
