% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front_diversity.R
\name{pareto_front}
\alias{pareto_front}
\title{pareto_front}
\usage{
pareto_front(X, sens)
}
\arguments{
\item{X}{a numerical matrix n x p of the set of point}

\item{sens}{a character vector of size p allowing only the two value in
c("min", "max"). A jth "min" value of sens means that the function
seeking the pareto front in order to minmize the jth component}
}
\value{
return a numerical matrix corresponding to the set of point belonging
to pareto front
}
\description{
Find the pareto front inside a set of point
}
