% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quant_reg_global_risk.R
\name{quant_reg_global_risk}
\alias{quant_reg_global_risk}
\title{quant_reg_global_risk}
\usage{
quant_reg_global_risk(
  X,
  Y,
  tau,
  allowed_dependence,
  initial_eta = 2.5,
  k = 0.2,
  floor = 10/NROW(X),
  max_iter = 20,
  tune = FALSE,
  path_tracking = NULL
)
}
\arguments{
\item{X}{data.frame of the dependent variables.}

\item{Y}{data.frame of the response variables.}

\item{tau}{vector of size NCOL(Y) containing the risk in ]0,1[ of each Y.}

\item{floor}{threshold to prevent estimate hard quantile with not enought data}

\item{max_iter}{integer to stop search if not converged. 20 by default.}

\item{path_tracking}{path where to write the step of the running function.}
}
\description{
quant_reg_global_risk
}
