% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsga_tunage.R
\name{sensibilite_nsga}
\alias{sensibilite_nsga}
\title{sensibilite_nsga}
\usage{
sensibilite_nsga(
  X,
  fn,
  sens = rep("min", length(fn)),
  N_MC = 10^4,
  B = 50,
  N = 50
)
}
\description{
sensibilite_nsga
}
