% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_selection.R
\name{tournament_selection}
\alias{tournament_selection}
\title{Tournement selection}
\usage{
tournament_selection(X, N)
}
\arguments{
\item{X}{a matrix/data.frame of all the selected individus evaluation}

\item{N}{size of the Qt population to generate}
}
\value{
a list of couple index which must be use for crossing over
}
\description{
Select a couple of individu (parents) to the crossing over
}
\examples{
sum(1:10)
}
