% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_operators.R
\name{crossing_over_uniforme}
\alias{crossing_over_uniforme}
\title{Crossing over operator}
\usage{
crossing_over_uniforme(parents, n_echange, param_mut)
}
\arguments{
\item{couple}{a matrix/data.frame of two individus}

\item{crossing_position}{a sorted vector of variable position to switch}
}
\value{
a matrix/data.frame of the two childs from the crossing over of the
parents
}
\description{
This function switch components for the given variables position between two
individus (parent) which have been selected after the tounament
}
\examples{
sum(1:10)
}
