% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint_definition.R
\name{def_cstr_X_space}
\alias{def_cstr_X_space}
\title{def_cstr_X_space}
\usage{
def_cstr_X_space(X, m = rep(10, NCOL(X)))
}
\description{
Find the space where points where observed. Send back a function to check if
a/several point(s) belong to the space where the points where observed
}
\details{
@param X matrix of dimension n x d observed points
 @param m a vector of size d to indicate how many reference points create for each dimension.
}
\examples{
#'library(tidyverse)

#generate data ---
#parameter of the data
n = 100
d = 2 #dim de X

mini = rep(0, d)
maxi = rep(10, d)


#uniform
X = sapply(seq_len(d), function(i){
  runif(n, mini[i], maxi[i])
})
plot(X)

#avec des trous
b = matrix(c(0, 0, 5, 5), byrow = T, ncol = d)
B = matrix(c(4, 4, 10, 10), byrow = T, ncol = d)
p = NROW(B) #nombre de "foyer"

X_c = sapply(seq_len(d), function(i){
  sapply(seq_len(p), function(k){
    runif(n/p, b[k,i], B[k,i])
  })
})
plot(X_c)

#use function
res = def_cstr_X_space(X_c)

#plot reference points with two colors depending or their feasability
plot(res$ref_points, pch=3, col = "orange")
points(res$feasible_ref_points, pch=3, col = "blue")

#check only one point
i=3
points(X[i,1], X[i,2], col = c("red", "green")[as.numeric(res$g(X = X[i,])) + 1])

#check several points
points(X, col = c("red", "green")[as.numeric(res$g(X)) + 1])

}
