% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsga.R
\name{NSGA}
\alias{NSGA}
\title{NSGA II algorithm}
\usage{
NSGA(
  X,
  fn,
  n_objective,
  sens = rep("min", n_objective),
  N = NROW(X),
  g = NULL,
  k_tournament,
  n_echange = 1,
  n_bloc = 2,
  crossing_over_method = "uniforme",
  mutation_method = "simple",
  freq_m = 0.2,
  type_var = NULL,
  distri_Xi = NULL,
  seed = NULL,
  TT = 50,
  verbose = TRUE,
  front_tracking = TRUE,
  updateProgress = NULL
)
}
\arguments{
\item{X}{a matrix/data.frame of an initial population}

\item{N}{the size of the population P before the generation of individus.
By default N=NROW(X)/2}

\item{ff}{a list of list containing the function or predicting model and
if the objectif function should be maximise "max" or minimize "min"}

\item{crossing_over_size}{number of variable to switch when doing the
crossing over}

\item{freq_mutation}{a vector of numerics value varying from 0 to 1 which
represent the frequency of mutation of each variable}
}
\value{
a vecteur of rank associated with the individus of X
}
\description{
The non sorting genetic algorithm is finding the Pareto front
}
\examples{
library(tidyverse)
set.seed(1234)
n <- 300
q = 3
mini <- rep(x = -2, times = q)
maxi <- rep(x = 2, times = q)

X <- lapply(seq_len(q), function(k){
  res <- data.frame(X=runif(n = n, min = mini[k], max = maxi[k]))
  names(res) <- paste0("X", k)
  res
}) \%>\% bind_cols()

fn <- list(
  Y1 = function(X) unlist(X[1] + 2*X[2] + X[3]),
  Y2 = function(X) unlist(-X[1] - X[2] - 3*X[3])
)

res = NSGA(X = X, fn, N=50)
plot(res$Y)

res = NSGA(X = X, fn, N=50, sens = rep("max", length(fn)))
plot(res$Y)

}
