% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsga.R
\name{NSGA}
\alias{NSGA}
\title{NSGA II algorithm}
\usage{
NSGA(X, ff, N, crossing_over_size, freq_mutation, seed, B = 500,
  verbose = TRUE)
}
\arguments{
\item{X}{a matrix/data.frame of an initial population}

\item{ff}{a list of list containing the function or predicting model and
if the objectif function should be maximise "max" or minimize "min"}

\item{N}{the size of the population P before the generation of individus.
By default N=NROW(X)/2}

\item{crossing_over_size}{number of variable to switch when doing the
crossing over}

\item{freq_mutation}{a vector of numerics value varying from 0 to 1 which
represent the frequency of mutation of each variable}
}
\value{
a vecteur of rank associated with the individus of X
}
\description{
The non sorting genetic algorithm is finding the Pareto front
}
\examples{
library(tidyverse)
set.seed(123)
n <- 300
q = 3
mini <- rep(x = -2, times = q)
maxi <- rep(x = 2, times = q)

X <- lapply(seq_len(q), function(k){
    res <- data.frame(X=runif(n = n, min = mini[k], max = maxi[k]))
    names(res) <- paste0("X", k)
    res
}) \%>\% bind_cols()

ff <- list(
    list(f = function(X) unlist(X[1] + 2*X[2] + X[3]),
         sens = "min"),
    list(f = function(X) unlist(-X[1] - X[2] - 3*X[3]),
         sens = 'min')
)

library(NSGA2)
NSGA(X=X[1:50,], ff, N = 50, crossing_over_size = 2, freq_mutation = rep(0.3, q),
    seed = 123, constraints = NULL, B=50, crossing_type = "hybrid",
    verbose = TRUE, time_out = 5)

}
