% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optisure.R
\name{optisure}
\alias{optisure}
\title{optisure}
\usage{
optisure(
  X,
  Y,
  sens = rep("max", NCOL(Y)),
  quantile_utility_idx = rep(TRUE, NCOL(Y)),
  g = NULL,
  X_space_csrt = TRUE,
  tau = rep(0.5, NCOL(Y)),
  globale_tau = rep(F, NCOL(Y)),
  alpha = 0.15,
  N = NULL,
  TT = NULL,
  seed = NULL,
  updateProgress = NULL,
  ...
)
}
\description{
optisure
}
\examples{
# generate data ----
library(tidyverse)
library(optisure)
set.seed(1234)
n <- 300

# X quanti ----
# X
d1 = 3
mini <- rep(x = -2, times = d1)
maxi <- rep(x = 2, times = d1)
X1 <- sapply(seq_len(d1), function(k){
  runif(n = n, min = mini[k], max = maxi[k])
}) \%>\% as.data.frame()
colnames(X1) = paste0("X", seq_len(d1))
X = X1

#calcul des Y selon une relation lineaire + du bruit
p = 2
fn = lapply(seq_len(p), function(j){
  beta = runif(n = d1, min = -2, max = 2)
  function(X) {
    as.matrix(X) \%*\% beta
  }
})
names(fn) = paste0("Y", seq_len(p))

Y = lapply(fn, function(f){
  f(X) + rnorm(n)
}) \%>\% as.data.frame(row.names = seq_len(n))

# Opti
res = optisure(X, Y, N=20, TT=10, alpha = 0.5)

plot(ellipse_var(Y$Y1, Y$Y2, alpha = 0.05), type = 'l', xlab = "Y1", ylab = "Y2",
     xlim = c(-8,8), ylim = c(-10,6))
points(Y, col = "grey")
points(res$Y, type = 'b', col = "blue", lwd = 3)

#X quali----
d2 = 2
lev = sample(x = 2:4, d2, replace = TRUE)
X2 <- sapply(seq_len(d2), function(k){
  as.factor(sample(x = seq_len(lev[k]), size = n, replace = TRUE))
}) \%>\% as.data.frame()
colnames(X2) = paste0("X", (d1+1):(d1+d2))

X = cbind(X1, X2)
head(X)

#calcul des Y selon une relation lineaire + du bruit
fn = lapply(seq_len(p), function(j){
  beta = runif(n = (d1 + sum(lev-1)), min = -2, max = 2)
  function(X) {
    is_num = sapply(X, is.numeric)
    X1 = X[, is_num]
    X2 = X[, !is_num]
    X2 = model.matrix(~., data = X2)[,-1]
    X = cbind(X1, X2)
    as.matrix(X) \%*\% beta
  }
})
names(fn) = paste0("Y", seq_len(p))

#opti
res = optisure(X, Y, N=20, TT=10, alpha = 0.5)

plot(ellipse_var(Y$Y1, Y$Y2, alpha = 0.05), type = 'l', xlab = "Y1", ylab = "Y2",
     xlim = c(-8,8), ylim = c(-10,6))
points(Y, col = "grey")
points(res$Y, type = 'b', col = "blue", lwd = 3)

}
